npc/pyproject.toml
[tool.poetry]
name = "npc"
version = "0.1.0"
packages = [
    { include = "npc", from = "src" }
]
description = ""
authors = ["Taylor Hearn <hearntaylor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.29.5"
torch = "2.0.0"
sentence-transformers = "^3.0.1"
config = "^0.5.1"
llama-index = "^0.11.3"
llama-index-llms-anthropic = "^0.3.0"
llama-index-embeddings-huggingface = "^0.3.1"
pydantic = "^2.9.0"
rich = "^13.8.1"
json-repair = "^0.29.5"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
EZGmail = "^2024.8.13"
PerplexiPy = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
